클래스 구조 변경
- 기존 클래스 CDungeonGenerator을 CVirtualCoordGenerator로 변경하고 AContentsGenerator 상속 해제
- CVirtualGenerator를 CGridGenerator의 자식으로 설정함으로써 CGridGenerator만이 접근 가능하도록 함
- CVirtualGenerator에서 생성된 가상 좌표를 기준으로 게임 오브젝트를 생성하는 클래스들이 AContentsGenerator 추상 클래스를 상속하고, CGridGenerator와는 별도로 존재
